Fix bug with labels containing URL delimiters
Due to a discrepancy between spec output and test expectaions a
conditional was added to fix this that wasn't part of the spec. It
still isn't clear which is correct; the specs or tests. This
conditional proved to be overzealous in trying to bridge the gap, so we
remove it in favor of the spec since it affects valid domains.